Allow export of used Family template instantiation
Identifiers for components and co are generated via template
instantiation. If you have two separate compilation units like a library
and an application using this library which pass a Registry around you
really want to have both use the same identifier or you are in for a lot
of undefined behavior fun.

Using the same compiler (version) for both is a good first step, but you
will also need the template instantiation symbols to have public
visibility in the library so that the application can use them instead
of generating its own instances (with different Identifiers).

On Windows symbols are by default hidden in dlls, on Linux the default
tends to be public, but is easily changed by -fvisibility=hidden and
recommended for a respectable library.

Note that you don't need nor possibly want to export the Family
instances in a library if EnTT isn't part of your public interface.