Intoduced a macro for suppressing MSVC warning C4307
The new macro can be used in user code for suppressing warnings when using
entt::hashed_string in constexpr context. The workaround provided in the FAQ

```c++
    #define HS(str)\
        __pragma(warning(push))\
        __pragma(warning(disable:4307))\
        entt::hashed_string{str}\
        __pragma(warning(pop))
    #define HS(str) entt::hashed_string{str}
```

can be reduced to

```c++
    #define HS(str) ENTT_HASHED_STRING_SUPPRESS_WARNING entt::hashed_string{str}
```

The new macro is also applied in the `ENTT_NAMED_TYPE` macro to reduce the
number of false warnings in user code. It is suggested that users should use
the macro also in their own specialisation of the `entt::named_type_traits`
template class.

The documentation was updated to mention a simplified form of warning
suppression and the newly added macro.

closes #348

Signed-off-by: Innokentiy Alaytsev <alaitsev@gmail.com>