Attempt to make header files for EnTT show up in IDE
I use QtCreator when programming on linux. When working with a project that consumes EnTT, if I navigate to the EnTT folder and tell the IDE to expand the EnTT target, it displays none of the headers.

This isn't so much an EnTT bug as it is an annoying cross-complication between CMake, the way CMake does header-only libs, QtCreator, and the specifics of the EnTT CMakeLists.txt file.

This change attempts to properly describe to any IDEs that the EnTT interface library has, as header files, anything under the "src" directory.

Whether a specific IDE will recognize this and properly display the headers for the EnTT target is up to that IDE.

Note that this also will cause a newly created headerfile, or a deleted header file, in the EnTT src/ folder to re-generate CMake, which I think is desirable, but others may not.